Set module_configuration instead of handler_configuration in DLH and TPHandler. Move appdalIssues.hpp to public header dir
